+++ lib/maru/builder.ex
@@ -60,8 +60,16 @@ defmodule Maru.Builder do
    |> List.flatten
 
     quote do
-      for i <- @endpoints |> Enum.reverse do
-        Module.eval_quoted __MODULE__, (i |> Endpoint.dispatch), [], __ENV__
+      @endpoints |> Enum.reverse |> Enum.map fn ep ->
+        Module.eval_quoted __MODULE__, (ep |> Endpoint.dispatch), [], __ENV__
+      end
+
+      @endpoints
+   |> Enum.group_by(fn ep -> {ep.version, ep.path} end)
+   |> Enum.map fn {{version, path}, endpoints} ->
+        unless Enum.any?(endpoints, fn i -> i.method == {:_, [], nil} end) do
+          Module.eval_quoted __MODULE__, (Endpoint.dispatch_405 version, path), [], __ENV__
+        end
       end
 
       defp endpoint(conn, _), do: conn