+++ test/builder/exceptions_test.exs
@@ -15,7 +15,7 @@ defmodule Maru.Builder.ExceptionsTest do
       end
 
       get :test3 do
-        raise Maru.Exceptions.NotFound, method: "GET"
+        raise "err"
       end
 
       rescue_from MatchError do
@@ -28,9 +28,14 @@ defmodule Maru.Builder.ExceptionsTest do
         "function error"
       end
 
+      rescue_from Maru.Exceptions.MethodNotAllow do
+        status 405
+        "MethodNotAllow"
+      end
+
       rescue_from :all, as: e do
         status 500
-        e.method
+        e.message
       end
     end
 
@@ -41,6 +46,9 @@ defmodule Maru.Builder.ExceptionsTest do
     assert %Plug.Conn{resp_body: "function error", status: 500} = RescueTest.call(conn2, [])
 
     conn3 = conn(:get, "test3")
-    assert %Plug.Conn{resp_body: "GET", status: 500} = RescueTest.call(conn3, [])
+    assert %Plug.Conn{resp_body: "err", status: 500} = RescueTest.call(conn3, [])
+
+    conn4 = conn(:post, "test3")
+    assert %Plug.Conn{resp_body: "MethodNotAllow", status: 405} = RescueTest.call(conn4, [])
   end
 end