+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * remove deprecated functions
   * add build\_embedded and start\_permanent options
   * add support to configure :port by system environment like {:system, "PORT"}
+  * add `match` DSL to handle all method
 
 ## v0.7.1 (2015-9-7)
 * Bugfix