+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 * Enhancements
   * remove deprecated functions
   * add build\_embedded and start\_permanent options
-  * add support to configure :port by system environment like {:system, "PORT"}
+  * Support configure :port by system environment like {:system, "PORT"}
   * add `match` DSL to handle all method
+  * Support HTTP 405 method not allowed
 
 ## v0.7.1 (2015-9-7)
 * Bugfix